requirements.txt
# Data Processing
pandas>=2.0.0
numpy>=1.24.0
PyMuPDF>=1.23.0  # PDF processing
lxml>=4.9.0       # XML processing
beautifulsoup4>=4.12.0

# Machine Learning
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# LLM & NLP
openai>=1.0.0
langchain>=0.1.0
tiktoken>=0.5.0
spacy>=3.7.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Jupyter & Visualization
jupyter>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Kaggle
kaggle>=1.5.0
